refactored native segwit address derivation
