+++ account/index.ts
@@ -90,8 +90,8 @@ export async function restoreWalletWithAccounts(
             id: index,
             stxAddress: response.stxAddress,
             btcAddress: response.btcAddress,
-            dlcBtcAddress: response.dlcBtcAddress,
-            dlcBtcPublicKey: response.dlcBtcPublicKey,
+            mainBtcAddress: response.mainBtcAddress,
+            mainBtcPublicKey: response.mainBtcPublicKey,
             ordinalsAddress: response.ordinalsAddress,
             masterPubKey: response.masterPubKey,
             stxPublicKey: response.stxPublicKey,
@@ -125,12 +125,12 @@ export async function createWalletAccount(
     stxAddress,
     btcAddress,
     ordinalsAddress,
-    dlcBtcAddress,
+    mainBtcAddress,
     masterPubKey,
     stxPublicKey,
     btcPublicKey,
     ordinalsPublicKey,
-    dlcBtcPublicKey,
+    mainBtcPublicKey,
   } = await walletFromSeedPhrase({
     mnemonic: seedPhrase,
     index: BigInt(accountIndex),
@@ -143,8 +143,8 @@ export async function createWalletAccount(
       id: accountIndex,
       stxAddress,
       btcAddress,
-      dlcBtcAddress,
-      dlcBtcPublicKey,
+      mainBtcAddress,
+      mainBtcPublicKey,
       ordinalsAddress,
       masterPubKey,
       stxPublicKey,
