+++ tests/mocks/restore.mock.ts
@@ -1,7 +1,9 @@
-import { Account } from "../../types";
+import { Account } from '../../types';
 
-export const testSeed = 'force kite borrow once shine pluck couch swift crystal swamp crumble essay';
+export const testSeed =
+  'force kite borrow once shine pluck couch swift crystal swamp crumble essay';
 
+  
 export const walletAccounts: Account[] = [
   {
     id: 0,
@@ -10,6 +12,8 @@ export const walletAccounts: Account[] = [
     btcPublicKey: '032215d812282c0792c8535c3702cca994f5e3da9cd8502c3e190d422f0066fdff',
     ordinalsAddress: 'bc1pr09enf3yc43cz8qh7xwaasuv3xzlgfttdr3wn0q2dy9frkhrpdtsk05jqq',
     ordinalsPublicKey: '5b21869d6643175e0530aeec51d265290d036384990ee60bf089b23ff6b9a367',
+    mainBtcAddress: 'bc1qf8njhm2nj48x9kltxvmc7vyl9cq7raukwg6mjk',
+    mainBtcPublicKey: '023537a32d5ab338a6ba52f13708ea45c1e3cb33c26aff3fa182d9c66fd4b636ff',
     stxAddress: 'SP147ST7ESA3RES888QQMV6AK7GZK93ZR74A0GM7V',
     stxPublicKey: '025df9b0ea2c81e4f8360bf9a16638ed3678bc84dbdc04124f5db86996999aa9a8',
   },
@@ -18,11 +22,35 @@ export const walletAccounts: Account[] = [
     masterPubKey: '024d30279814a0e609534af1d1969b7c24a6918029e1f9cb2134a427ebfb1f17c3',
     btcAddress: '3EMRvkWMLaUfzHPA7Un5qfLZDvbXHn385u',
     btcPublicKey: '022e633aba8838c039b2d2214f51ed284d3da7f585744f8975606376c23483d2c1',
+    mainBtcAddress: 'bc1qgxk5z97gempeav0xdrdcezhn9att7dev6gavyc',
+    mainBtcPublicKey: '03af366875dc848b7d8306b13c872f448791b7f7fbc8948e4452a07a74bff5f035',
     ordinalsAddress: 'bc1pnc669rz0hyncjzxdgfeqm0dfhfr84ues4dwgq4lr47zpltzvekss4ptlxw',
     ordinalsPublicKey: '380447c41546e736f3d4bf9dc075d2301f5252f33156e3564fd393eeffdaa347',
     stxAddress: 'SP1BKESAFFV8ACW007HACXB93VHRFHP83BT24Z3NF',
     stxPublicKey: '0302ec9c40f8d5daf319bf4b1556c7f51f1eb449dd96d05e7ed42a1056451dd656',
   },
+  {
+    id: 2,
+    masterPubKey: '024d30279814a0e609534af1d1969b7c24a6918029e1f9cb2134a427ebfb1f17c3',
+    btcAddress: '3DUdrysnoh9CvMdtWhavTqRXhFvzg3aKNQ',
+    btcPublicKey: '036b7ca83c456a5b577d5631cd38c76c2d0540c25e6787bb4394001aebaea42d20',
+    mainBtcAddress: 'bc1q3xkz9q3fsxt6pxjxaq5w7yf6mhh4zxcu5d2fev',
+    mainBtcPublicKey: '03eb31409ecdcf37688b7d7ce1e915d8660208704695298b701a5652b05856d879',
+    ordinalsAddress: 'bc1pgu2hx6vkp0yja2sdeugkzs3xlxdwnj0wrasqletv8v98cn0vntesa2h8zv',
+    ordinalsPublicKey: 'ee3ad253e6696962b43ca3dd83a40dc3184e11ff0679f972a84704be36bc9225',
+    stxAddress: 'SP2BCTPVVHWN2VA8N85GRXJAPS10NR5YNFC6SY5SR',
+    stxPublicKey: '0268c588454446d214cbf7c7b0e970ba8cb3425f8976a045eed7f98112db7e1942',
+  },
+  {
+    btcAddress: '35xwHGuXcn4ufM2KDttL2w5PPsT9DRSoR6',
+    btcPublicKey: '03289a3a4d3b28565f5b5459d1570d8954bc08d09fb8d276154ff9cfeb0447bd10',
+    id: 3,
+    mainBtcAddress: 'bc1q85l49tvgpg09ztcn6dnp0ph4pxqy82xzhq48w2',
+    mainBtcPublicKey: '037465ac1d27643f071b88203497ba0d07bdb05e060923cd93a32d765b4f5f7158',
+    masterPubKey: '024d30279814a0e609534af1d1969b7c24a6918029e1f9cb2134a427ebfb1f17c3',
+    ordinalsAddress: 'bc1p6cz06dkrngdeuapgfuvahpnkvelmrcryk422kmd0y7zmuu2nechsyan7nk',
+    ordinalsPublicKey: '2cf69ea437dd236422e4490244c94e97664f4ad065dd280d83df5d5a22bef6b9',
+    stxAddress: 'SP1X4G7MZNTHGFF6DA58PN7EWQ72DGMF7834P6X8',
+    stxPublicKey: '0289e90678355f7b73e53a798fdb0a656d8faa2a186900f64cb8c06b7155f0e419',
+  },
 ];
-
-
