+++ types/wallet.ts
@@ -1,8 +1,8 @@
 export interface BaseWallet {
   stxAddress: string;
   btcAddress: string;
-  dlcBtcAddress: string;
-  dlcBtcPublicKey: string;
+  mainBtcAddress: string;
+  mainBtcPublicKey: string;
   ordinalsAddress: string;
   masterPubKey: string;
   stxPublicKey: string;
